Remove meaningless failed tests
Until we have removed cartodb, it's hard to even make these tests work,
since in the app, we load cartodb from a CDN, and karma doesn't seem to
be a big fan of that.

Once cartodb is gone, we can include leaflet properly and maybe have
some tests that use it.

for TSD-46