Parallelize locations download across two requests
Join the requests client-side once they arrive

In testing, this resulted in faster download (close to 2x), but a much longer time
waiting for the server, bringing the total time to worse than a single request.

On very slow connections, this technique should still be faster, since the download time
would dominate. If there is a way to get CKAN to process two simultaneous requests
without killing performance, maybe this can still win in every case.