Hack around a bug in chrome/v8's JS optimizer
We were getting
`Uncaught TypeError: Cannot read property 'slice' of undefined`
from the `_zoomEnd` handler used internally by leaflet when the map was
zoomed by any means, about 50% of the time.

The impact was rendering errors and semi-broken interactivity of the
map. Zoom still worked but click-drag to pan did not.

Through debugging, it was determined that the cause is some issue in
the javascript compiler's optimizer. Placing a `debugger;` statement
would prevent the bug from occurring, but doing so also disables the
optimizer for that function. The optimizer-bug hypothesis was validated
by adding other code known to disable optimization of javascript
functions: an empty `try {} catch(e) {}` made the bug go away, as did a
`with({});` statement anywhere in the function.

Finally, through experimentation it was discovered that simply moving
the erroring expression into an inline anonymous function also made the
bug go away. This is the fix contained in this commit, which
monkey-patches itself into leaflet if it detects itself running on
chrome.

It would be nice to reduce the original bug to a minimal setup that
exhibits the issue in order to prepare a bug report for the chrome
team, but this has yet to be completed.