Split `selected` for schools and polygons
Turns out we need to have separated variables for them, since we have to keep a concept of the
currently-selected district when transitioning back to schools view.

Some code ends up being nearly-duplicated, but it is also much simpler in many places now, since we don't have
to switch on `viewMode` any more.