Use JSONP in oldIE to avoid XDomain issues
It's a hack, but I can't even find a clear answer as to whether or not cross-domain requests are supported in IE9 by angular in the first
place.

I tried to mess around with it to make it work. It's possible some server headers are screwing it up (content-type, maybe?)

But in any case, all we are doing is boring old GETs. JSONP is awful but hey, it works, it's only used for old ie, ... I can sleep at
night.